Type graph edges and renamer memberships with Link
The incidence graph stored edge kinds as u8 and cast Link back and
forth: incidence and propagation cast on the way in, the fast renamer
decoded kinds with value-equality guards, and the slow renamer encoded
its three memberships as 0, 1, and 2 on its own. Graph edges, edge
hashing, and both renamers now carry Link. Link derives Ord in
discriminant order, so every sort and hash is unchanged, and the
memberships are the Particle, Holder, and Owner links a token already
has from its own side.

The refinement depth shared by the incremental network and the state
signature is now propagation::DEPTH instead of 4, 5, and 3 kept in
agreement by hand. Benchmark reports keep identical events, work, and
statistics.

+++ language/test/scheduling.rs
@@ -411,13 +411,10 @@ fn capture() {
             .edge
             .iter()
             .flatten()
-            .filter(|&&(kind, _)| kind == crate::link::Link::Capture as u8)
+            .filter(|&&(kind, _)| kind == crate::link::Link::Capture)
             .copied()
             .collect::<Vec<_>>();
-        assert_eq!(
-            capture,
-            vec![(crate::link::Link::Capture as u8, graph.frame[1])]
-        );
+        assert_eq!(capture, vec![(crate::link::Link::Capture, graph.frame[1])]);
         let renamed = state.rename(&[1, 0], &[0, 1]);
         assert_eq!(renamed.resource.len(), 1);
         assert_eq!(renamed.resource[&42], 0);
